Introduce AbstractQueryState class
